+++ pycommon_database/audit_mongo.py
@@ -66,31 +66,36 @@ def _common_audit(model, base):
 
 
 def _versioning_audit(model, base):
-    class AuditModel(CRUDModel, base=base, table_name=f'audit_{model.__tablename__}', audit=False):
+    class AuditModel(CRUDModel, base=base, table_name='audit', audit=False):
         """
-        Class providing Audit fields for a MONGODB model.
+        Class providing the audit for all versioned MONGODB models.
         """
+        table_name = Column(str, is_primary_key=True)
+        revision = Column(int, is_primary_key=True, index_type=IndexType.Unique)
+
         audit_user = Column(str, is_primary_key=True)
         audit_date_utc = Column(datetime.datetime, is_primary_key=True, index_type=IndexType.Unique)
         audit_action = Column(Action)
 
         @classmethod
-        def audit_add(cls, audit_date: datetime.datetime):
-            cls._audit_action(action=Action.Insert, audit_date=audit_date)
+        def audit_add(cls, revision: int):
+            cls._audit_action(action=Action.Insert, revision=revision)
 
         @classmethod
-        def audit_update(cls, audit_date: datetime.datetime):
-            cls._audit_action(action=Action.Update, audit_date=audit_date)
+        def audit_update(cls, revision: int):
+            cls._audit_action(action=Action.Update, revision=revision)
 
         @classmethod
-        def audit_remove(cls, audit_date: datetime.datetime):
-            cls._audit_action(action=Action.Delete, audit_date=audit_date)
+        def audit_remove(cls, revision: int):
+            cls._audit_action(action=Action.Delete, revision=revision)
 
         @classmethod
-        def _audit_action(cls, action: Action, audit_date: datetime.datetime):
+        def _audit_action(cls, action: Action, revision: int):
             cls.__collection__.insert_one({
+                cls.table_name: model.__tablename__,
+                cls.revision.name: revision,
                 cls.audit_user.name: current_user_name(),
-                cls.audit_date_utc.name: audit_date,
+                cls.audit_date_utc.name: datetime.datetime.utcnow(),
                 cls.audit_action.name: action.value,
             })